Feat: 전체 검색 페이지 동작 구현, info 팝업 추가 (#131)
